removing old detector infrastructure
