+++ apps/common/include/apps/common/display_volumes_rz.inl
@@ -6,7 +6,7 @@
  */
 #pragma once
 
-#include "core/proto_detector.hpp"
+#include "core/detector.hpp"
 #include "core/transform_store.hpp"
 #include "io/csv_io.hpp"
 #include "utils/containers.hpp"
