+++ tests/CMakeLists.txt
@@ -1,7 +1,6 @@
 enable_testing()
 
 add_subdirectory(common)
-add_subdirectory(io)
 
 if(DETRAY_UNIT_TESTS)
     add_subdirectory(unit_tests)
