+++ tests/benchmarks/CMakeLists.txt
@@ -1,10 +1,12 @@
+message(STATUS "Testing: 'detray::benchmarks' enabled")
+
 set (BENCHMARK_ENABLE_INSTALL OFF)
 set (BENCHMARK_ENABLE_TESTING OFF)
 add_subdirectory("${PROJECT_SOURCE_DIR}/extern/benchmark" "extern/benchmark")
 
 macro(add_detray_benchmark TESTNAME FILES PLUGIN_LIBRARY)
     add_executable(${TESTNAME} ${FILES})
-    target_link_libraries(${TESTNAME} benchmark detray::core detray::tests_common detray::tests_io detray::io ${PLUGIN_LIBRARY})
+    target_link_libraries(${TESTNAME} benchmark detray::core detray::tests_common detray::io ${PLUGIN_LIBRARY})
     install(TARGETS ${TESTNAME} RUNTIME DESTINATION bin)
 endmacro()
