+++ tests/unit_tests/CMakeLists.txt
@@ -1,4 +1,4 @@
-message(STATUS "Testing: 'detray' unit tests")
+message(STATUS "Testing: 'detray::unit_tests' enabled")
 
 mark_as_advanced(
     BUILD_GMOCK BUILD_GTEST BUILD_SHARED_LIBS
@@ -11,7 +11,7 @@ include(GoogleTest)
 macro(add_detray_test TESTNAME FILES PLUGIN_LIBRARY)
     add_executable(${TESTNAME} ${FILES})
     target_link_libraries(${TESTNAME} PRIVATE gtest gmock gtest_main)
-    target_link_libraries(${TESTNAME} PRIVATE detray::core detray::tests_common detray::tests_io)
+    target_link_libraries(${TESTNAME} PRIVATE detray::core detray::tests_common)
     target_link_libraries(${TESTNAME} PRIVATE  ${PLUGIN_LIBRARY} ${UNIT_TEST_EXTRA_LIBRARIES})
     set_target_properties(${TESTNAME} PROPERTIES FOLDER tests)
     add_test(unit_test_${TESTNAME} ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${TESTNAME})
@@ -22,8 +22,7 @@ endmacro()
 set(all_unit_tests "plugin;core")
 list(APPEND all_unit_tests "annulus2;cylinder3;rectangle2;ring2;single3;unmasked;trapezoid2")
 list(APPEND all_unit_tests "cylinder_intersection;planar_intersection;intersection_kernel")
-list(APPEND all_unit_tests "detector;single_layer_components")
-list(APPEND all_unit_tests "transform_store;masks_container")
+list(APPEND all_unit_tests "detector;transform_store;masks_container")
 
 add_subdirectory(core)
 if(DETRAY_ARRAY_PLUGIN)
