--- tests/unit_tests/array/array_single_layer_components.cpp
@@ -1,9 +1,0 @@
-/** Detray library, part of the ACTS project (R&D line)
- * 
- * (c) 2020 CERN for the benefit of the ACTS project
- * 
- * Mozilla Public License Version 2.0
- */
-
-#include "plugins/array_defs.hpp"
-#include "tests/common/test_single_layer_components.inl"
